E6EW9S4 move the log's tests out of the player's file, and share the helpers both need
