+++ source/test/e2e-gui/theatre.pw.ts
@@ -1,34 +1,5 @@
-import type {Page} from '@playwright/test';
 import {expect, test} from './fixtures.js';
-
-// The suite shares one server, and a board left parked in the past never
-// finishes loading for the next test.
-const returnToLive = async (page: Page) => {
-	const exit = page.getByTestId('theatre-exit');
-	// The button carries the word "Resume" only while the board is in the past,
-	// so its absence is what live looks like.
-	const resume = page.getByRole('button', {name: 'Resume', exact: true});
-
-	if ((await exit.count()) > 0) await exit.click();
-
-	// Pressed until it takes, rather than asked once: a request for live is a
-	// message like any other and a socket replaced under it drops it silently,
-	// which would leave the board parked in the past for the rest of the file.
-	await expect(async () => {
-		if ((await resume.count()) === 0) return;
-		// Bounded: the word goes the moment live lands, and an unbounded click on
-		// a button that has already lost its name would sit out the whole retry
-		// budget.
-		await resume.click({timeout: 1_000});
-
-		await expect(resume).toHaveCount(0, {timeout: 2_000});
-	}).toPass({timeout: 20_000});
-};
-
-const openBoard = async (page: Page, appUrl: string) => {
-	await page.goto(appUrl);
-	await expect(page.getByTestId('board-switcher')).toContainText('Default');
-};
+import {openBoard, returnToLive} from './live-board.js';
 
 test('the play button opens a player over a dimmed board', async ({
 	page,
@@ -172,171 +143,6 @@ test('the transport pauses and resumes, and the movie ends on a full bar', async
 	await expect(toggle).toHaveAttribute('aria-label', 'Play again');
 
 	await returnToLive(page);
-	expect(pageErrors).toEqual([]);
-});
-
-// The crawl is a slice of the script, not a list grown as events land, which
-// is what keeps a long movie from adding a node per event to the overlay.
-test('the pop-out puts the log beside the board', async ({
-	page,
-	appUrl,
-	pageErrors,
-}) => {
-	await openBoard(page, appUrl);
-	await page.getByTestId('theatre-play').click();
-
-	const log = page.getByTestId('event-log');
-	await expect(log).toHaveCount(0);
-
-	const toggle = page.getByTestId('theatre-log-toggle');
-	await toggle.click();
-	await expect(log).toBeVisible();
-	await expect(toggle).toHaveAttribute('aria-pressed', 'true');
-
-	// Lines arrive as the movie plays. How many the panel holds at most is
-	// event-log.test.ts's job — a bound asserted here would only be a bound on
-	// the seeded window, which is smaller than the cap and so proves nothing.
-	const lines = page.getByTestId('log-line');
-	await expect.poll(async () => await lines.count()).toBeGreaterThan(1);
-
-	// The day each run of lines belongs to is called once above them, rather
-	// than repeated on every line.
-	await expect(page.getByTestId('log-day').first()).toBeVisible();
-
-	// Every line is marked with its kind, commits included. The dot is a
-	// pseudo-element rather than a node of its own — the panel holds hundreds of
-	// these — so what is asserted is the colour each row hands it.
-	// Evaluated as a string, like the other DOM reads in this suite: these files
-	// are type-checked against the Node libs, which have no `HTMLElement`.
-	const dotColours = (await page.evaluate(
-		`[...document.querySelectorAll('[data-testid="log-line"]')]` +
-			`.map(row => row.style.getPropertyValue('--epiq-log-dot').trim())`,
-	)) as string[];
-
-	expect(dotColours).toHaveLength(await lines.count());
-	expect(dotColours.every(colour => colour.length > 0)).toBe(true);
-
-	// A panel, not a wash over the board: it takes its own width and the first
-	// swimlane starts to the right of where it ends.
-	const logBox = (await log.boundingBox())!;
-	const lane = (await page
-		.getByTestId('swimlane-handle')
-		.first()
-		.boundingBox())!;
-
-	expect(logBox.width).toBeGreaterThan(100);
-	expect(lane.x).toBeGreaterThanOrEqual(logBox.x + logBox.width);
-
-	await toggle.click();
-	await expect(log).toHaveCount(0);
-
-	await returnToLive(page);
-	expect(pageErrors).toEqual([]);
-});
-
-// The panel is the board's, not the movie's: the pop-out and the Log box are
-// two controls over one flag, and leaving the player does not take it away.
-test('the log stays on the board after the player leaves', async ({
-	page,
-	appUrl,
-	pageErrors,
-}) => {
-	await openBoard(page, appUrl);
-
-	const log = page.getByTestId('event-log');
-	const box = page.getByTestId('log-toggle');
-
-	await expect(log).toHaveCount(0);
-	await box.click();
-	await expect(log).toBeVisible();
-
-	// Live, with no movie anywhere near it: the tail of the window.
-	await expect
-		.poll(async () => await page.getByTestId('log-line').count())
-		.toBeGreaterThan(0);
-
-	await page.getByTestId('theatre-play').click();
-	await expect(page.getByTestId('theatre-player')).toBeVisible();
-	await expect(log).toBeVisible();
-
-	await page.getByTestId('theatre-exit').click();
-	await expect(page.getByTestId('theatre-player')).toHaveCount(0);
-
-	// Still there, and still the same panel.
-	await expect(log).toBeVisible();
-
-	await box.click();
-	await expect(log).toHaveCount(0);
-
-	await returnToLive(page);
-	expect(pageErrors).toEqual([]);
-});
-
-// The log is drawn from the scrubber's window, which is fetched when the window
-// moves — not when the board changes. Without the log asking for it too, a
-// swimlane made while the panel is open never reaches the panel listing it.
-test('the log picks up what happens on the board while it is open', async ({
-	page,
-	appUrl,
-	pageErrors,
-}) => {
-	await openBoard(page, appUrl);
-
-	const box = page.getByTestId('log-toggle');
-	await box.click();
-	await expect(page.getByTestId('event-log')).toBeVisible();
-
-	const lines = page.getByTestId('log-line');
-	await expect.poll(async () => await lines.count()).toBeGreaterThan(0);
-
-	// A swimlane, because it writes an event without needing a ticket first.
-	const name = `log-live-${Date.now()}`;
-	await page.getByTestId('add-swimlane').click();
-	await page.getByPlaceholder('swimlane name').fill(name);
-	await page.getByPlaceholder('swimlane name').press('Enter');
-
-	await expect(page.getByText(name, {exact: true}).first()).toBeVisible();
-
-	// Its line arrives without the window being touched. Asserted by what the
-	// line says rather than by the count going up: the log is capped, and on a
-	// board that has already filled it a new line pushes the oldest off the top
-	// instead of adding to the tally.
-	await expect(page.getByTestId('event-log')).toContainText(name);
-
-	await box.click();
-	await expect(page.getByTestId('event-log')).toHaveCount(0);
-	expect(pageErrors).toEqual([]);
-});
-
-// Folding is what bounds the panel: a folded day is one row however many
-// events it holds, so a long history is a handful of rows until asked for.
-test('a day folds to its divider and opens again', async ({
-	page,
-	appUrl,
-	pageErrors,
-}) => {
-	await openBoard(page, appUrl);
-
-	const box = page.getByTestId('log-toggle');
-	await box.click();
-
-	const lines = page.getByTestId('log-line');
-	await expect.poll(async () => await lines.count()).toBeGreaterThan(0);
-
-	// The newest day is the one open by default.
-	const day = page.getByTestId('log-day').last();
-	await expect(day).toHaveAttribute('aria-expanded', 'true');
-
-	await day.click();
-	await expect(day).toHaveAttribute('aria-expanded', 'false');
-	await expect(lines).toHaveCount(0);
-
-	await day.click();
-	await expect(day).toHaveAttribute('aria-expanded', 'true');
-	await expect.poll(async () => await lines.count()).toBeGreaterThan(0);
-
-	await box.click();
-	await expect(page.getByTestId('event-log')).toHaveCount(0);
 	expect(pageErrors).toEqual([]);
 });
